+++ test/lib/testcommon.js
@@ -45,16 +45,16 @@ function execPlus(args, cb) {
         args.log.trace({exec: true, command: command, execOpts: execOpts,
             err: err, stdout: stdout, stderr: stderr}, 'exec done');
         if (err) {
-            cb(
-                new VError(err,
+            var niceErr = new VError(err,
                     '%s:\n'
                     + '\tcommand: %s\n'
                     + '\texit status: %s\n'
                     + '\tstdout:\n%s\n'
                     + '\tstderr:\n%s',
                     args.errMsg || 'exec error', command, err.code,
-                    stdout.trim(), stderr.trim()),
-                stdout, stderr);
+                    stdout.trim(), stderr.trim());
+            niceErr.code = err.code;
+            cb(niceErr, stdout, stderr);
         } else {
             cb(null, stdout, stderr);
         }